requirements.txt
torch>=2.0.0
diffusers>=0.20.0
transformers>=4.30.0
accelerate>=0.20.3
safetensors
sentence-transformers
scikit-learn
faiss-cpu
streamlit
Pillow
numpy
pandas
matplotlib
opencv-python
tqdm
torchvision==0.18.*